utils_vlm/requirements.txt
opencv-python>=4.8.0
torch>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
Pillow>=10.0.0
python-dotenv>=1.0.0
openai>=1.0.0
segmentation-models-pytorch>=0.3.3
pytorch-lightning>=2.0.0
tqdm>=4.65.0
